# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "JobPilot API",
    "description": "Local JSON API for JobPilot search and application management.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/queue": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Queue",
        "description": "Return ranked daily queue items.",
        "operationId": "get_queue_api_queue_get",
        "parameters": [
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QueueItem"
                  },
                  "title": "Response Get Queue Api Queue Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Detail",
        "description": "Return complete job details including analysis, score, and drafts.",
        "operationId": "get_job_detail_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/prepare": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Prepare Job",
        "description": "Prepare application draft materials for a job.",
        "operationId": "prepare_job_api_jobs__job_id__prepare_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrepareResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/mark": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Mark Job",
        "description": "Transition a job's status and record application metadata.",
        "operationId": "mark_job_api_jobs__job_id__mark_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkJobRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusUpdate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/skip": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Skip Job",
        "description": "Mark a job as skipped.",
        "operationId": "skip_job_api_jobs__job_id__skip_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusUpdate"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/jd": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Add Jd",
        "description": "Attach full job description text to a job.",
        "operationId": "add_jd_api_jobs__job_id__jd_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddJDRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddJDResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Applications",
        "description": "List applications optionally filtered by job status.",
        "operationId": "list_applications_api_applications_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationItem"
                  },
                  "title": "Response List Applications Api Applications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/follow-ups": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Follow Ups",
        "description": "List applied jobs where follow-up is due.",
        "operationId": "get_follow_ups_api_follow_ups_get",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FollowUpItem"
                  },
                  "title": "Response Get Follow Ups Api Follow Ups Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/answers/ask": {
      "post": {
        "tags": [
          "answers"
        ],
        "summary": "Ask Question",
        "description": "Generate or retrieve an answer to an application question.",
        "operationId": "ask_question_api_answers_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/answers/{answer_id}": {
      "put": {
        "tags": [
          "answers"
        ],
        "summary": "Update Answer",
        "description": "Supply or update an answer for an answer bank item.",
        "operationId": "update_answer_api_answers__answer_id__put",
        "parameters": [
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Answer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionSuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/answers/{answer_id}/approve": {
      "post": {
        "tags": [
          "answers"
        ],
        "summary": "Approve Answer",
        "description": "Mark an answer bank entry as approved.",
        "operationId": "approve_answer_api_answers__answer_id__approve_post",
        "parameters": [
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Answer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionSuccessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddJDRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 200000,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "AddJDRequest"
      },
      "AddJDResult": {
        "properties": {
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "char_count"
        ],
        "title": "AddJDResult"
      },
...
      "ApplicationItem": {
        "properties": {
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "applied_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "referral_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Contact"
          },
          "prepared_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prepared At"
          },
          "tailored_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tailored Summary"
          },
          "short_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Note"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "company",
          "title",
          "status"
        ],
        "title": "ApplicationItem"
      },
...
      "AskQuestionRequest": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AskQuestionRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "fingerprint": {
            "type": "string",
            "title": "Fingerprint"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "remote_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remote Type"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "apply_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apply Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snippet"
          },
          "posted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "discovered_at": {
            "type": "string",
            "title": "Discovered At"
          },
          "last_seen_at": {
            "type": "string",
            "title": "Last Seen At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobAnalysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "score": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScoreDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "application": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApplicationDraft"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "fingerprint",
          "source",
          "company",
          "title",
          "url",
          "discovered_at",
          "last_seen_at",
          "status"
        ],
        "title": "JobDetail"
      },
      "JobStatusUpdate": {
        "properties": {
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "previous_status": {
            "type": "string",
            "title": "Previous Status"
          },
          "new_status": {
            "type": "string",
            "title": "New Status"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "referral_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Contact"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "previous_status",
          "new_status",
          "timestamp"
        ],
        "title": "JobStatusUpdate"
      },
      "MarkJobRequest": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "referral_contact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referral Contact"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "MarkJobRequest"
      },
...
      "PrepareResult": {
        "properties": {
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "bullet_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bullet Ids"
          },
          "outreach_draft": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outreach Draft"
          },
          "used_fallback": {
            "type": "boolean",
            "title": "Used Fallback",
            "default": false
          },
          "validation_errors": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Errors"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "tier",
          "summary",
          "note"
        ],
        "title": "PrepareResult",
        "description": "Outcome of preparing one job."
      },
...
      "QueueItem": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank"
          },
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "matched_skills": {
            "type": "string",
            "title": "Matched Skills"
          },
          "gaps": {
            "type": "string",
            "title": "Gaps"
          },
          "flags": {
            "type": "string",
            "title": "Flags"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "prepared": {
            "type": "string",
            "title": "Prepared"
          }
        },
        "type": "object",
        "required": [
          "rank",
          "job_id",
          "company",
          "title",
          "location",
          "tier",
          "score",
          "matched_skills",
          "gaps",
          "flags",
          "url",
          "prepared"
        ],
        "title": "QueueItem",
        "description": "Single item in the daily ranked queue."
      },
...
    }
  }
}